I think default handlers are a good idea
