+++ .gitignore
@@ -7,6 +7,7 @@
 /pkg/
 /spec/reports/
 /tmp/
+/log/
 spec/dummy/
 .ruby-version
 .env*
