+++ lib/ship_quiet_logistics/handlers.rb
@@ -1,0 +1,2 @@
+require 'ship_quiet_logistics/handlers/error_message_handler'
+require 'ship_quiet_logistics/handlers/process_shipment_handler'
